Split out preference utilities
